+++ src/algorithms/data_type_adapter/adapters/ibyte_to_cbyte.cc
@@ -37,7 +37,7 @@ IbyteToCbyte::IbyteToCbyte(const ConfigurationInterface* configuration,
 {
     const std::string default_input_item_type("byte");
     const std::string default_output_item_type("lv_8sc_t");
-    const std::string default_dump_filename("../data/input_filter.dat");
+    const std::string default_dump_filename("./data_type_adapter.dat");
 
     DLOG(INFO) << "role " << role_;