+++ src/algorithms/input_filter/adapters/beamformer_filter.cc
@@ -34,7 +34,7 @@ BeamformerFilter::BeamformerFilter(
       dump_(configuration->property(role + ".dump", false))
 {
     const std::string default_item_type("gr_complex");
-    const std::string default_dump_file("./data/input_filter.dat");
+    const std::string default_dump_file("./input_filter.dat");
     item_type_ = configuration->property(role + ".item_type", default_item_type);
     dump_filename_ = configuration->property(role + ".dump_filename", default_dump_file);
     DLOG(INFO) << "role " << role_;