+++ src/algorithms/input_filter/adapters/notch_filter_lite.cc
@@ -37,7 +37,7 @@ NotchFilterLite::NotchFilterLite(const ConfigurationInterface* configuration,
       dump_(configuration->property(role + ".dump", false))
 {
     const std::string default_item_type("gr_complex");
-    const std::string default_dump_file("./data/input_filter.dat");
+    const std::string default_dump_file("./input_filter.dat");
     const float default_p_c_factor = 0.9;
     const float default_pfa = 0.001;
     const float default_samp_freq = 4000000;