+++ tests/unit-tests/signal-processing-blocks/acquisition/galileo_e1_pcps_quicksync_ambiguous_acquisition_gsoc2014_test.cc
@@ -306,7 +306,7 @@ void GalileoE1PcpsQuickSyncAmbiguousAcquisitionGSoC2014Test::config_2()
 
     /*Unset this flag to eliminates data logging for the Validation of results
     probabilities test*/
-    dump_test_results = true;
+    dump_test_results = false;
 
     num_of_realizations = 100;
 
@@ -882,7 +882,7 @@ TEST_F(GalileoE1PcpsQuickSyncAmbiguousAcquisitionGSoC2014Test, ValidationOfResul
                         {
                             std::stringstream filenamepd;
                             filenamepd.str("");
-                            filenamepd << "../data/test_statistics_" << gnss_synchro.System
+                            filenamepd << "./test_statistics_" << gnss_synchro.System
                                        << "_" << gnss_synchro.Signal << "_sat_"
 #if USE_GLOG_AND_GFLAGS
                                        << gnss_synchro.PRN << "CN0_dB_0_" << FLAGS_e1_value_CN0_dB_0 << "_dBHz.csv";
@@ -908,7 +908,7 @@ TEST_F(GalileoE1PcpsQuickSyncAmbiguousAcquisitionGSoC2014Test, ValidationOfResul
                         {
                             std::stringstream filenamepf;
                             filenamepf.str("");
-                            filenamepf << "../data/test_statistics_" << gnss_synchro.System
+                            filenamepf << "./test_statistics_" << gnss_synchro.System
                                        << "_" << gnss_synchro.Signal << "_sat_"
 #if USE_GLOG_AND_GFLAGS
                                        << gnss_synchro.PRN << "CN0_dB_0_" << FLAGS_e1_value_CN0_dB_0 << "_dBHz.csv";